code_shot/requirements.txt
Django>=6.0
Pillow>=10.0.0
Pygments>=2.15.0